requirements.txt
Flask
python-dotenv
pydantic
psycopg2-binary
PyJWT